requirements.txt
fastapi==0.128.8
uvicorn==0.40.0
pydantic==2.12.5
requests==2.32.5
python-dotenv==1.2.1
sentence-transformers==5.2.2
torch==2.10.0
chromadb==1.5.0